requirements.txt
requests
python-dotenv
langchain_openai
langchain-core
langgraph
opencv-python
mss
sounddevice 
numpy
websockets
fastapi
google-genai
pdfminer
python-multipart
pdfminer.six
pyaudio
pydub
uvicorn[standard]